pyproject.toml
[project]
name = "markdoc-py"
version = "0.1.0"
description = "Python port of Markdoc"
authors = [
    { name = "Oskar Skoog", email = "oskar@osd.se" },
]
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
    "markdown-it-py>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pdoc>=16.0.0",
    "pytest>=9.0.2",
    "ruff>=0.1.0",
    "tox>=4.0.0",
]

[project.scripts]
markdoc-py-docs = "markdocpy._docs:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.ruff]
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["markdocpy"]
